+++ clients/ctl/admin-ui/cypress/e2e/datasets-classify.cy.ts
@@ -1,4 +1,9 @@
-import { stubPlusHealth } from "../support/stubs";
+import { stubDatasetCrud, stubPlus } from "cypress/support/stubs";
+
+import {
+  ClassifyInstance,
+  ClassifyStatusEnum,
+} from "~/features/common/plus.slice";
 
 /**
  * This test suite is a parallel of datasets.cy.ts for testing Dataset features when the user has
@@ -7,7 +12,11 @@ import { stubPlusHealth } from "../support/stubs";
  */
 describe("Datasets with Fides Classify", () => {
   beforeEach(() => {
-    stubPlusHealth();
+    stubDatasetCrud();
+    stubPlus(true);
+    cy.intercept("GET", "/api/v1/plus/classify", {
+      fixture: "classify/list.json",
+    }).as("getClassifyList");
   });
 
   describe("Creating datasets", () => {
@@ -36,15 +45,9 @@ describe("Datasets with Fides Classify", () => {
       cy.intercept("POST", "/api/v1/generate", {
         fixture: "generate/dataset.json",
       }).as("postGenerate");
-      cy.intercept("POST", "/api/v1/dataset", { fixture: "dataset.json" }).as(
-        "postDataset"
-      );
       cy.intercept("POST", "/api/v1/plus/classify", {
         fixture: "classify/create.json",
       }).as("postClassify");
-      cy.intercept("GET", "/api/v1/dataset", { fixture: "datasets.json" }).as(
-        "getDatasets"
-      );
 
       // Confirm the request.
       cy.getByTestId("confirmation-modal").getByTestId("continue-btn").click();
@@ -65,15 +68,6 @@ describe("Datasets with Fides Classify", () => {
   });
 
   describe("List of datasets with classifications", () => {
-    beforeEach(() => {
-      cy.intercept("GET", "/api/v1/dataset", { fixture: "datasets.json" }).as(
-        "getDatasets"
-      );
-      cy.intercept("GET", "/api/v1/plus/classify", {
-        fixture: "classify/list.json",
-      }).as("getClassifyList");
-    });
-
     it("Shows the each dataset's classify status", () => {
       cy.visit("/dataset");
       cy.wait("@getDatasets");
@@ -97,13 +91,6 @@ describe("Datasets with Fides Classify", () => {
       cy.intercept("GET", "/api/v1/dataset/*", {
         fixture: "classify/dataset-in-review.json",
       }).as("getDataset");
-      cy.intercept("GET", "/api/v1/data_category", {
-        fixture: "data_categories.json",
-      }).as("getDataCategory");
-
-      cy.intercept("GET", "/api/v1/plus/classify", {
-        fixture: "classify/list.json",
-      }).as("getClassifyList");
     });
 
     /**
@@ -153,6 +140,57 @@ describe("Datasets with Fides Classify", () => {
         identifiability: "Unlinked Pseudonymized",
         taxonomyEntities: ["system.operations"],
       });
+    });
+  });
+
+  describe("Approve classification button", () => {
+    beforeEach(() => {
+      cy.intercept("GET", "/api/v1/dataset/*", {
+        fixture: "classify/dataset-in-review.json",
+      }).as("getDataset");
+
+      cy.intercept("PUT", "/api/v1/dataset/*", {
+        fixture: "classify/dataset-in-review.json",
+      }).as("putDataset");
+      cy.intercept("PUT", "/api/v1/plus/classify", {
+        fixture: "classify/update.json",
+      }).as("putClassify");
+    });
+
+    it("Updates blank fields with top classifications", () => {
+      cy.visit("/dataset/dataset_in_review");
+      cy.wait("@getDataset");
+      cy.wait("@getClassifyList");
+
+      // The button triggers a chain of requests that will modify the classify list response.
+      // TODO(#1120): When the APIs are locked in, the list request will be replaced with a
+      // get-by-fides-key request, and the PUT requests can have assertions about their content.
+      cy.fixture("classify/list.json").then((instances) => {
+        const instance: ClassifyInstance = instances.find(
+          (ci: ClassifyInstance) => ci.id === "classification_in_review"
+        );
+        instance.datasets[0].status = ClassifyStatusEnum.REVIEWED;
+
+        cy.intercept("GET", "/api/v1/plus/classify", {
+          body: instances,
+        }).as("getClassifyList");
+      });
+
+      cy.getByTestId("approve-classification-btn")
+        .should("be.enabled")
+        .click()
+        .should("be.disabled");
+
+      // The mutations should run.
+      cy.wait("@putDataset");
+      cy.wait("@putClassify");
+
+      // The instances should be re-fetched.
+      cy.wait("@getClassifyList");
+
+      // The updated status should make the button disappear.
+      cy.getByTestId("approve-classification-btn").should("not.exist");
+      cy.getByTestId("toast-success-msg");
     });
   });
 });