+++ src/fidesctl/api/ctl/sql_models.py
@@ -85,6 +85,48 @@ class PGEncryptedString(TypeDecorator):
         pass
 
 
+class ClassificationDetail(Base):
+    """
+    The SQL model for a classification instance
+    """
+
+    __tablename__ = "cls_classification_detail"
+    instance_id = Column(String(255))
+    status = Column(Text)
+    dataset = Column(Text)
+    collection = Column(Text)
+    field = Column(Text)
+    labels = Column(JSON)
+    created_at = Column(DateTime(timezone=True), server_default=func.now())
+    updated_at = Column(
+        DateTime(timezone=True),
+        server_default=func.now(),
+        onupdate=func.now(),
+    )
+    # get the details from a classification json output (likely aggregate and options etc.)
+
+
+class ClassificationInstance(Base):
+    """
+    The SQL model for a classification instance
+    """
+
+    __tablename__ = "cls_classification_instance"
+
+    status = Column(Text)
+    organization_key = Column(Text)
+    dataset_key = Column(Text)
+    dataset_name = Column(Text)
+    target = Column(Text)
+    type = Column(Text)
+    created_at = Column(DateTime(timezone=True), server_default=func.now())
+    updated_at = Column(
+        DateTime(timezone=True),
+        server_default=func.now(),
+        onupdate=func.now(),
+    )
+
+
 # Privacy Types
 class DataCategory(Base, FidesBase):
     """